chore: extract `pkg/imager/quirks` to `pkg/machinery`
To make it possible to use it without pulling the whole Talos.

(cherry picked from commit 3dd1f4e88c22734f03f7609791558b8bbbae3756)
