+++ pkg/machinery/imager/quirks/quirks.go
